requirements.txt
dspy-ai
openai
pydantic
pydantic-settings
sentence-transformers
pymilvus